separate scanner activity
