+++ app/build.gradle
@@ -27,10 +27,10 @@ repositories {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:21.0.3'
-    compile 'com.android.support:support-v4:21.0.3'
     //compile 'com.google.zxing:core:2.2'
     //compile 'com.embarkmobile:zxing-android-minimal:1.2.1@aar'
+    compile 'com.android.support:appcompat-v7:21.0.3'
+    compile 'com.android.support:support-v4:21.0.3'
     compile 'me.dm7.barcodescanner:zxing:1.8.4'
-    compile 'com.android.support.constraint:constraint-layout:+'
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'
 }
